+++ Velocity/Heterodyne/Heterodyne/src/I2C.h
@@ -109,7 +109,7 @@ class I2C
     uint8_t read(int, int);
     uint8_t read(uint8_t, uint8_t, uint8_t);
     uint8_t read(int, int, int);
-    uint8_t read(uint8_t, uint8_t, uint8_t*);
+    uint8_t read(uint8_t, uint16_t, uint8_t*);
     uint8_t read(uint8_t, uint8_t, uint8_t, uint8_t*);
